refactor(mod)!: rename the go module of this repository
We now consider this project a hard fork, meaning we do not seek to
integrate the changes in the original repository as those a too
structural and only present for how specific requirements (e.g.
determinism, execution control flow, etc...).

By changing the module name we do not rely on importing the original
module and use the replace directore anymore, to be used this project
must now be imported directly.